requirements.txt
requests==2.28.2
beautifulsoup==4.11.2
lxml==4.9.2
pandas
matplotlib
seaborn
Pillow=